requirements.txt
pytesseract==0.3.10
langchain==0.1.13
python-dotenv==1.0.1
langchain-openai==0.1.1
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9